requirements.txt
agno==1.1.8
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.8
databases==0.9.0
dataclasses-json==0.6.7
deepseek==1.0.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.5.0
ecdsa==0.19.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gradio==5.15.0
gradio_client==1.7.0
gspread==5.12.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.6
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
lancedb==0.20.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.3.6
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.2
oauthlib==3.2.2
openai==1.61.1
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.14.0
propcache==0.2.1
psycopg==3.2.5
psycopg2==2.9.10
pyarrow==19.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pylance==0.23.2
pypdf==5.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
semantic-version==2.10.0
sentence-transformers==3.4.1
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.1
tantivy==0.22.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.1.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
